Split tests by method
