+++ .gitignore
@@ -2,3 +2,4 @@ node_modules
 index.js
 index.d.ts
 *.map
+!tests/*
